+++ benchmarks/benchmarker/utils.py
@@ -7,17 +7,257 @@ import base64
 import json
 import logging
 import os
+import signal
 import struct
 import subprocess
+import sys
+import threading
 import time
+from collections.abc import Generator, Iterator
+from contextlib import contextmanager
+from pathlib import Path
 
 import requests as requests_lib
 
 logger = logging.getLogger(__name__)
 
+STARTUP_TIMEOUT = 600
+REPO_ROOT = Path(__file__).resolve().parents[2]
+GPU_CLEANUP_SCRIPT = REPO_ROOT / ".github/scripts/ensure_gpus_idle.sh"
+GPU_IDLE_THRESHOLD_MB = 2048
+GPU_IDLE_WAIT_SECONDS = 600
+GPU_IDLE_POLL_SECONDS = 5
 WAV_HEADER_SIZE = 44
 SSE_DATA_PREFIX = "data: "
 SSE_DONE_MARKER = "data: [DONE]"
+
+
+@contextmanager
+def disable_proxy() -> Generator[None, None, None]:
+    """Temporarily disable proxy env vars for loopback requests."""
+    proxy_vars = (
+        "HTTP_PROXY",
+        "HTTPS_PROXY",
+        "http_proxy",
+        "https_proxy",
+        "ALL_PROXY",
+        "all_proxy",
+        "NO_PROXY",
+        "no_proxy",
+    )
+    saved_env = {k: os.environ[k] for k in proxy_vars if k in os.environ}
+    for k in proxy_vars:
+        os.environ.pop(k, None)
+    try:
+        yield
+    finally:
+        for k in proxy_vars:
+            os.environ.pop(k, None)
+        os.environ.update(saved_env)
+
+
+def no_proxy_env() -> dict[str, str]:
+    """Return a copy of os.environ with proxy variables removed, for subprocess use."""
+    proxy_keys = {"http_proxy", "https_proxy", "all_proxy", "no_proxy"}
+    return {k: v for k, v in os.environ.items() if k.lower() not in proxy_keys}
+
+
+def server_log_file(tmp_path_factory, prefix: str = "server_logs") -> Path | None:
+    """Capture server logs to a file on CI; stream to the terminal locally."""
+    is_ci = os.environ.get("GITHUB_ACTIONS") == "true"
+    if not is_ci:
+        return None
+    return tmp_path_factory.mktemp(prefix) / "server.log"
+
+
+def stop_server(proc: subprocess.Popen) -> None:
+    """Gracefully stop the server process group, tolerating already-dead processes."""
+    try:
+        os.killpg(os.getpgid(proc.pid), signal.SIGTERM)
+        proc.wait(timeout=30)
+    except (ProcessLookupError, ChildProcessError):
+        return
+    except subprocess.TimeoutExpired:
+        try:
+            os.killpg(os.getpgid(proc.pid), signal.SIGKILL)
+            proc.wait(timeout=10)
+        except (ProcessLookupError, ChildProcessError):
+            return
+
+
+def wait_for_gpu_memory_release(
+    *,
+    memory_threshold_mb: int | None = None,
+    wait_timeout_seconds: int | None = None,
+    poll_seconds: int | None = None,
+) -> None:
+    """Kill orphan GPU processes and block until every GPU is below threshold."""
+    if not GPU_CLEANUP_SCRIPT.exists():
+        raise FileNotFoundError(f"GPU cleanup script missing: {GPU_CLEANUP_SCRIPT}")
+
+    env = os.environ.copy()
+    env["OMNI_CI_GPU_MEMORY_CLEAN_THRESHOLD_MB"] = str(
+        memory_threshold_mb
+        if memory_threshold_mb is not None
+        else GPU_IDLE_THRESHOLD_MB
+    )
+    env["OMNI_CI_GPU_CLEAN_WAIT_SECONDS"] = str(
+        wait_timeout_seconds
+        if wait_timeout_seconds is not None
+        else GPU_IDLE_WAIT_SECONDS
+    )
+    env["OMNI_CI_GPU_CLEAN_POLL_SECONDS"] = str(
+        poll_seconds if poll_seconds is not None else GPU_IDLE_POLL_SECONDS
+    )
+
+    print(
+        f"[gpu cleanup] running ensure_gpus_idle "
+        f"(threshold={env['OMNI_CI_GPU_MEMORY_CLEAN_THRESHOLD_MB']} MiB)...",
+        flush=True,
+    )
+    result = subprocess.run(
+        ["bash", str(GPU_CLEANUP_SCRIPT)],
+        env=env,
+        check=False,
+    )
+    if result.returncode != 0:
+        raise RuntimeError(
+            "GPU memory was not released after stopping the inference server. "
+            f"ensure_gpus_idle.sh exit={result.returncode}"
+        )
+
+
+def wait_healthy(
+    proc: subprocess.Popen,
+    port: int,
+    log_file: Path | None,
+    timeout: int = STARTUP_TIMEOUT,
+) -> None:
+    """Wait for a server to report healthy, stopping it and raising on failure."""
+    try:
+        with disable_proxy():
+            wait_for_service(
+                f"http://localhost:{port}",
+                timeout=timeout,
+                server_process=proc,
+                server_log_file=log_file,
+                health_body_contains="healthy",
+            )
+    except Exception as exc:
+        stop_server(proc)
+        log_text = (
+            log_file.read_text() if log_file is not None and log_file.exists() else ""
+        )
+        message = str(exc)
+        if log_text and log_text not in message:
+            message = f"{message}\n{log_text}"
+        if isinstance(exc, TimeoutError):
+            raise TimeoutError(message) from exc
+        if isinstance(exc, RuntimeError):
+            raise RuntimeError(message) from exc
+        raise
+
+
+def start_server_from_cmd(
+    cmd: list[str],
+    log_file: Path | None,
+    port: int,
+    timeout: int = STARTUP_TIMEOUT,
+    env: dict[str, str] | None = None,
+    tee: bool = False,
+) -> subprocess.Popen:
+    """Start a server from an arbitrary command and wait until healthy."""
+    process_env = os.environ.copy()
+    if env is not None:
+        process_env.update(env)
+    if log_file is None:
+        proc = subprocess.Popen(
+            cmd,
+            env=process_env,
+            start_new_session=True,
+        )
+    elif tee:
+        # Tee (file + stdout): TP=2 fixture wants the file for grep + live
+        # output for `pytest -s`. Pattern from sglang's popen_launch_server.
+        log_handle = open(log_file, "w")
+        try:
+            proc = subprocess.Popen(
+                cmd,
+                env=process_env,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.STDOUT,
+                start_new_session=True,
+                text=True,
+                bufsize=1,
+            )
+        except Exception:
+            log_handle.close()
+            raise
+
+        def _tee_stdout(src, sink) -> None:
+            try:
+                for line in iter(src.readline, ""):
+                    sink.write(line)
+                    sink.flush()
+                    sys.stdout.write(line)
+                    sys.stdout.flush()
+            finally:
+                src.close()
+                sink.close()
+
+        # log_handle ownership is handed to the thread; its finally closes it.
+        threading.Thread(
+            target=_tee_stdout,
+            args=(proc.stdout, log_handle),
+            daemon=True,
+        ).start()
+    else:
+        with open(log_file, "w") as log_handle:
+            proc = subprocess.Popen(
+                cmd,
+                env=process_env,
+                stdout=log_handle,
+                stderr=subprocess.STDOUT,
+                start_new_session=True,
+            )
+    wait_healthy(proc, port, log_file, timeout=timeout)
+    return proc
+
+
+@contextmanager
+def managed_omni_server(
+    *,
+    model_path: str,
+    port: int,
+    host: str,
+    log_file: Path | None,
+    timeout: int = STARTUP_TIMEOUT,
+    wait_for_gpu_release: bool = True,
+) -> Iterator[None]:
+    """Start an ``sglang_omni.cli serve`` process and clean it up on exit."""
+    cmd = [
+        sys.executable,
+        "-m",
+        "sglang_omni.cli",
+        "serve",
+        "--model-path",
+        model_path,
+        "--port",
+        str(port),
+        "--host",
+        host,
+    ]
+    logger.info(f"Starting server: {' '.join(cmd)}")
+    if log_file is not None:
+        log_file.parent.mkdir(parents=True, exist_ok=True)
+    proc = start_server_from_cmd(cmd, log_file, port, timeout=timeout)
+    try:
+        yield
+    finally:
+        logger.info(f"Stopping server ({model_path})")
+        stop_server(proc)
+        if wait_for_gpu_release:
+            wait_for_gpu_memory_release()
 
 
 def get_wav_duration(wav_bytes: bytes) -> float:
